Break out arrow into common PercentageChangeArrowTile
